Provide an additional macro expansion for F128 in stdlib/tst-strtod.h
The macro F128 in stdlib/tst-strtod.h is defined to provide the literal
suffix for _Float128 constants.  It uses the macro __f128 (), which is
defined in bits/floatn.h to provide the correct literal suffix depending on
what is provided by the compiler.

However, F128 was not being expanded and only worked correctly, when
compiling with GCC 7 (or greater), since F128 is the literal suffix itself.
This patch adds an additional macro expansion so that the macro F128
expands to the correct literal suffix on older compilers.

	* stdlib/tst-strtod.h (MMFUNC): New macro to provide an addition
	macro expansion.
	(GEN_TEST_STRTOD_FOREACH): Use MMFUNC for _Float128.